+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Add static bool ServerSettingsModel.IsPrivateServer to allow clients to select the private LanguageForge
+
 ## [5.0.0] - 2021-10-28
 
 ### Fixed